Refactor Snitch frontend
Split into multiple smaller components.
